Changed exit from sprite to a cell, changed mapdata to dict
Exit is now cell 9 instead of a sprite. This meant a bunch of other
changes.

Map data is now stored in a coord-indexed dict instead of a list of
lists.

Also added some more error checking during level parsing.

Cell tests in the level class are still pretty messy. Will probably just
bite the bullet and make a class for each cell.