Changed sprite speed calculation, limited frame rate
Now sprite speed is set as the number of milliseconds it takes to travel
one cell (lower = faster).

Also capped framerate at 60 fps.

Plus, changed the way laserbeams update while a mirror is being moved in
or out of the beam; they are now stopped by a mirror between cells.
There is probably still a better way to do it.