Added Chip's Challenge level 2, ability to change cells, new cell hook
Levels now have a set_cell method which allow the changing of a cell.

Cells now have a player_inside() hook which runs when the player arrives
entirely in a cell.

These are used in CC level 2. which introduces the dirt block, which
turns water to mud. Chip stepping on mud turns it to floor.