Cells now render their own tiles; optimized redraw
Cell rendering is now handled inside the cell object instead of the main
game object. This allows cells to override the rendering process.

This has been done with the ice cell, which can blit a corner tile on
top of the ice tile.

Also, if a cell changes, just that cell is redrawn instead of the entire
background.