New level format using Tiled Map Editor
This uses a third-party program, Tiled, which outputs map data in a
couple of different formats, wrapped in XML. Currently using CSV, though
might switch to base64 or some other binary format.

This necessitates a second tileset, for use in building maps. Also
included is Tiled's tileset format which is based on the new tileset
image.

Also, for that reason, the Sentry now has its own tile.

And, switched the order of data being passed to the sprite generator. No
big deal.