Window calls the resize hook on resize, instead of on every frame
It does this by looping through a screen stack. Also, the hook gets
called when initializing the screen.

Now all that logic can go right into the subclasses, and almost nothing
is needed in the superclass.

I know I keep moving this stuff around, but it's pretty trial and error.