Add configuration file
Now default configuration is set in a single file instead of in the
constructors of each object, and as default arguments.

Had to extend the ConfigParser class because it lacked a few features
like parsing lists. Also it just reads a single section from the config,
so I don't have to pass the name of the section to every frigging
get call.